moved global constants into common/
